feat: move aws functions into dedicated module
